Refactor extension code structure and improve API handling
- Updated content.js to streamline message handling and improve readability.
- Refactored apiRequest.ts to utilize the Headers API for setting request headers.
- Removed backend.ts and integrated its functionality into backEndApi.ts.
- Enhanced type definitions in index.ts to include optional token for participants.
- Updated uiHandlers.ts to import BackEndApi from the new location.
- Introduced apiRoutes.js and backEndApi.js to manage API endpoints and requests.
- Implemented configurationManager.js to handle configuration retrieval from storage.
- Added utility functions for date formatting and current URL retrieval.
- Developed meetingHubClient.js to manage SignalR connections for real-time communication.
- Created chatPanel.js to manage UI updates for chat interactions.
- Implemented state management through stateManager.js to track application state.
- Enhanced user experience with toast notifications and tooltips.
- Improved overall code organization and modularity for better maintainability.
