+++ modules/extension/src/api/types.ts
@@ -1,0 +1,7 @@
+import {Participant} from "../core/types";
+
+export interface RequestAddBotModel {
+  meetId: string;
+  botName: string;
+  user: Participant;
+}
